fix: addressing PR review comments
